Update RotaryEncoder.h
change the _positionExt rangs  from 0-8091 to  0-65535

+++ RotaryEncoder.h
@@ -35,7 +35,7 @@ private:
   
   int8_t _oldState;
   
-  int _position;     // Internal position (4 times _positionExt)
+  long _position;     // Internal position (4 times _positionExt)
   int _positionExt;  // External position
 };
